Added run-time task configuration and support for each village to configure individual task.
Previously, we have each tasks configuration stored in memory, and in order to update it a recompile was required, now configuration is moved a json file, which can be updated at runtime.
Additionally, updated Tasks to be able to configured for a single villageId only, this allows fine level of configuration for each villages purposes.